+++ androidApp/build.gradle.kts
@@ -31,7 +31,7 @@ android {
         applicationId = "com.piappstudio.biggboss"
         minSdk = (findProperty("android.minSdk") as String).toInt()
         targetSdk = (findProperty("android.targetSdk") as String).toInt()
-        versionCode = 1
+        versionCode = 4
         versionName = "1.0"
     }
     compileOptions {
@@ -40,5 +40,14 @@ android {
     }
     kotlin {
         jvmToolchain(17)
+    }
+    buildTypes {
+        debug {
+            proguardFiles(getDefaultProguardFile("proguard-android-optimize.txt"), "proguard-rules.pro")
+        }
+        release {
+            isMinifyEnabled = true
+            proguardFiles(getDefaultProguardFile("proguard-android-optimize.txt"), "proguard-rules.pro")
+        }
     }
 }
