+++ shared/src/commonMain/resources/MR/base/strings.xml
@@ -9,4 +9,5 @@
     <string name= "title_nomination">Nominations</string>
     <string name= "title_others">Other</string>
     <string name = "title_eliminated">Eliminated</string>
+    <string name = "title_notification">Notification</string>
 </resources>
